minlp-tokenizer/requirements.txt
tensorflow>=1.14
pyahocorasick
regex

